更新清空购物车功能，将用户 ID 类型从 uint 修改为 uint32

+++ app/cart/biz/service/empty_cart.go
@@ -21,6 +21,6 @@ func (s *EmptyCartService) Run(req *cart.EmptyCartReq) (resp *cart.EmptyCartResp
 	if req.UserId ==0{
 		return 
 	}
-	err=model.EmptyCart(s.ctx,mysql.DB,uint(req.UserId))
+	err=model.EmptyCart(s.ctx,mysql.DB,req.UserId)
 	return
 }
